ISSUE #2243 - move rule/meta functions out of groups
